Fix error in multimodal plan
